backend/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
google-generativeai
scikit-learn
numpy
PyPDF2